sim: a scenario message stamp was still float, and clang is what said so
Cross-compiler determinism had never seen the fixed-point work: check-determinism
only golden-hashes the detmath shim standalone, not the sim, and the sim is where
this run put everything. So the sim itself was built with clang and run against
a gcc referee.

Clang refused to compile it. `Scenario::Msg::t` is the tick a message was queued
and stayed float when clock_ became integer, so `pending_.push_back({clock_, ...})`
narrows int32 to float in an initializer list -- ill-formed, rejected by clang,
accepted silently by gcc. Same class as the five other tick leftovers this run,
caught by a different compiler rather than by reading.

WHAT THE CROSS-BUILD RUN SAYS, which is the part worth having:

  gcc server + gcc client    hash=8f42b8ce04e60a86   (control)
  gcc server + CLANG client  hash=8f42b8ce04e60a86   0 desyncs

Byte-identical state after 1800 ticks between binaries from two different
compilers. That is the property the whole fixed-point conversion exists to buy,
and until now it had only been asserted for the transcendental shim.

The full suite also passes built with clang -- 15 of 16 binaries, yield_test not
being a target in that configuration.

check-determinism itself is still green: gcc and clang at -O0/-O2/-O3 all agree
on golden ab1ef54ae324bd0e. Its aarch64 legs SKIP here for want of a cross
toolchain, so ARM parity remains covered only by the Apple-silicon CI leg.

+++ src/sim/scenario.h
@@ -43,7 +43,9 @@ public:
 
     // Cosmetic display messages queued since the last drain (game time, target
     // player, text). Target -1 = All Players. NOT part of the hash.
-    struct Msg { float t = 0; int player = -1; std::string text; };
+    // `t` is the TICK the message was queued, matching Scenario::clock_. It was
+    // float, and clang rejected the narrowing that gcc accepted in silence.
+    struct Msg { int32_t t = 0; int player = -1; std::string text; };
     std::vector<Msg> drainMessages();
 
     bool showClock() const { return showClock_; }   // "Display gameclock" latched
